Don't dump the literal table
The literal table was dumped, so future reference to symbols could be
found. However, now we have proper packages and we don't need the
literal table anymore. Symbols will be interned when required.

We still use the literal table for conses and shorter references to
symbols during bootstraping.

This changes also makes the output file smaller.