Remove global value from *toplevel-compilations*
This change reveals a new bug. The `bool` codegen macro invokes
`convert`, but the codegen executes out of the context of the
compilations.

It means that CONVERT could call TOPLEVEL-COMPILATION to intern
T/NIL. The compiler will cache the result in the *LITERAL-TABLE*, but
the toplevel form was never included in the output.