Optimize function calls
Represent unbound symbol-function with JS function

Symbols are created with a fvalue property initialized to
`unboundFunction`. This function will throw a 'undefined function' error
when called.

The lisp function `symbol-function` checks if .fvalue is identical to
unboundFunction, so it will signal an error if unbound without having to
call the function.