Don't allow (fdefinition '(function foo))
Pretty sure that's not a good thing.  SBCL signals an error and the CLHS
defines function-name as

> A symbol or a list (setf symbol).

We extend  that for FFI with  list (oget …) also,  but (function symbol)
isn't really a thing …

It's kinda recursive, really, since you'd expect (function symbol) to be
essentially a macro version of (symbol-function (quote symbol)) which in
turn is (fdefinition (quote symbol)) …

Any rate, I've talked myself into destroying it.