Keep patch to ANSI-Test here.
The current patch is from @jackdaniel and will probably make it into the
Git version soon enough, but for the  moment, I don't expect there to be
“enough” things that we  need to patch in the test  suite itself to want
to maintain a fork.

More significantly, for things that are  so badly broken that we have to
patch the test suite (eg: perhaps  a reader macro bug or something), the
thing to do is  to fix JSCL; in the mean time, this  patch file can live
as a  “pebble in the  sandal” so that  disabled, broken tests  are never
forgotten completely.