♯J improvements; use reader macro function in JSCL reader
In  this patch,  moved the  CL (compat.lisp)  ♯J reader  into read.lisp,
where it can now be used in both models (SBCL/CL:READ or JSCL/CL:READ).

(NOTE:  Using  actual sharp-sign  ♯  here  so  Git  will not  treat  the
octothorpe number-sign as a comment marker.)

The following  is the revised  documentation on the  WITH-SHARP-J macro,
which  is the  best description  right now.  Note the  new comma/at-sign
forms described below.

————————————————————

The ♯J macro for reading JavaScript forms:

♯J:a:b:c  is equivalent  to (JSCL/FFI:OGET*  *ROOT* \"a\"  \"b\" \"c\"),
which     in     turn     is    equivalent     to     the     JavaScript
window[\"a\"][\"b\"][\"c\"]  or (when  the keys  are valid  identifiers)
also the same as window.a.b.c and a.b.c.

Note  that the  ♯J: sequence  is \emph{case-sensitive}  and begins  with
a colon  (:) after  the ♯J.  Each :-delimited substring  is the  name of
a key on the preceding object in the sequence.

TODO: An equivalent form where the  reference object does not have to be
*ROOT*  or where  identifiers  may  not have  to  be  strings. There  is
a discussion in http://github.com/jscl-project/jscl/Issues/

An experimental version is incorporated here. In this notation:

Begin with “♯J”. Follow this with :  to refer to the root object (normally
“window”). Follow it instead with @ to NOT refer to the root object, but
instead,  name   an  object  (locally   in  scope)  to   be  referenced.
For example, ♯J@foo is identical to FOO.

Each  subsequent section  is  introduced  by either  a  colon or  comma.
A  section introduced  by  : is  treated  as a  string  name. A  section
introduced by  a , is instead  the name of a  locally available (current
package) symbol, which is bound to the string value to be used as a key.

Examples:

♯J:foo:bar → window[\"foo\"][\"bar\"]
♯J@foo:bar → foo[\"bar\"]
♯J:foo,bar → window[\"foo\"][ bar ]
♯J@foo,bar → foo[ bar ]

Note that @  in the first position BOTH suppresses  the automatic use of
the root  object (window) AND  ALSO acts  like a comma,  “unquoting” the
following argument.

Note that the @ and , options  are EXPERIMENTAL and local to the Romance
Ⅱbranch. The discussion upstream may substantially change the meaning of
these options.

Rationale: @  and , are used  in Lisp as quasi-quote  operators, and are
not  valid JavaScript  identifier components,  so they  are unlikely  to
occur in either language in normal use.