Move HASH-TABLE to JSCL/CL; use storage-vectors
The storage-vector implementation here is  really just to take advantage
of the  KIND field, and  still stuffs everything  into a JS  Object; ie,
this is  a (VECTOR 1 JS:Object)  with a KIND property  attached. To make
the KIND more meaningful, it's stored always as a symbol (not a function
directly); ie,  (make-hash-table :test #'eql) is  converted backwards to
yield  a kind  of (HASH-TABLE  EQL)  — the  symbol (QUOTE  EQL) not  the
function-object (FUNCTION EQL).

EQUALP is still a FIXME/TODO, for now.