Fix docstring and body parsing
Consider cases where there are multiple declarations or a docstring in
between. Those cases do not seem very clearly specified in the CLHS,
but I have tried to interpret it in the more flexible way.

It fixes some code examples from PAIP and it is consistent with SBCL.

Some test cases were added.