Treat numbers as strings when accessing properties
With this change, numbers are now treated as strings when setting
properties of javascript objects. This aligns with how object properties
are accessed in javascript already:

```sh
$ node
Welcome to Node.js v18.7.0.
Type ".help" for more information.
> var o = {}
undefined
> o[111] = 222
222
> o["111"]
222
>
```

Before this, the following error would be thrown.

```
CL-USER> (jscl::oset 222 (jscl::new) 111)
ERROR: activechars.join is not a function
CL-USER>
```